+++ src/main/java/com/chungbazi/server/domain/policy/application/HomePolicyService.java
@@ -6,6 +6,8 @@ import com.chungbazi.server.domain.policy.application.cursor.PolicyCursor;
 import com.chungbazi.server.domain.policy.application.cursor.PolicyCursorParser;
 import com.chungbazi.server.domain.policy.application.cursor.RecentViewedPolicyCursor;
 import com.chungbazi.server.domain.policy.application.cursor.RecentViewedPolicyCursorParser;
+import com.chungbazi.server.domain.policy.application.mapper.HomePolicyResponseAssembler;
+import com.chungbazi.server.domain.policy.application.mapper.PolicyListResponseAssembler;
 import com.chungbazi.server.domain.policy.domain.entity.Policy;
 import com.chungbazi.server.domain.policy.domain.entity.RecentViewedPolicy;
 import com.chungbazi.server.domain.policy.domain.repository.RecentViewedPolicyRepository;
